// src/app/utils/models/adsense.model.ts
export interface PromotedBrandAd {
    id: string;
    pictureUrl: string;
    landingPageUrl: string;
    categoryName: string;
}

export interface VideoAdUnit {
    id: string;
    videoName: string;
    videoUrl: string;
    externalVideoLink: string;
}


export interface BannerAdUnit {
    id: string;
    pictureUrl: string;
    landingPageUrl: string;
    bannerType: string;
}